requirements.txt
numpy==1.24.1
torch==2.2.2 --index-url  https://download.pytorch.org/whl/cu121
torchvision==0.17.2 --index-url  https://download.pytorch.org/whl/cu121
timm==0.9.16
wandb==0.16.6
lightning==2.1.0
pytorch_optimizer==2.6.1
pytorch_forecasting==1.0.0
transformers==4.40.1
diffusers==0.21.3
accelerate==0.28.0
omegaconf
einops==0.3.2
pytorch-lightning==1.5.0
bchlib==0.7
torch_dct==0.1.6
lpips==0.1.4
facenet_pytorch==2.5.3
kornia==0.7.0
open_clip_torch==2.7.0
pytorch_msssim==1.0.0
compressai==1.2.4
tensorflow==2.9.0
torcheval
piqa==1.3.2
ninja==1.11.1.1
huggingface_hub==0.25.0
